took a rest from coding, watched some course materials to renew knowledge about Clean Architecture and decided to remake architecture in a different way. In this case all schema was reconsidered and files moved in different directions. All other work would have been done soon.
