+++ pom.xml
@@ -94,9 +94,9 @@
                         <id>default-cli</id>
                         <configuration>
                             <mainClass>com.example.mapsdemo/mapsdemo.HelloApplication</mainClass>
-                            <launcher>app</launcher>
-                            <jlinkZipName>app</jlinkZipName>
-                            <jlinkImageName>app</jlinkImageName>
+                            <launcher>APP</launcher>
+                            <jlinkZipName>APP</jlinkZipName>
+                            <jlinkImageName>APP</jlinkImageName>
                             <noManPages>true</noManPages>
                             <stripDebug>true</stripDebug>
                             <noHeaderFiles>true</noHeaderFiles>
