+++ src/APP/main.java
@@ -1,10 +1,10 @@
-package app;
+package APP;
 
 import javafx.application.Application;
 import javafx.scene.Scene;
 import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
-import view.LoadMapView;
+import VIEW.LoadMapView;
 
 public class main extends Application {
