+++ src/DATA_ACCESS/Database.java
@@ -1,4 +1,4 @@
-package database;
+package DATA_ACCESS;
 
 import java.sql.*;
