+++ src/USE_CASE/LoadEventsOutputBoundary.java
@@ -1,0 +1,4 @@
+package USE_CASE;
+
+public interface LoadEventsOutputBoundary {
+}
