+++ src/USE_CASE/LoadMapInputBoundary.java
@@ -1,0 +1,4 @@
+package USE_CASE;
+
+public interface LoadMapInputBoundary {
+}
