+++ src/USE_CASE/LoadMapInteractor.java
@@ -1,0 +1,4 @@
+package USE_CASE;
+
+public class LoadMapInteractor {
+}
