+++ src/USE_CASE/LoadMapOutputBoundary.java
@@ -1,0 +1,4 @@
+package USE_CASE;
+
+public interface LoadMapOutputBoundary {
+}
