+++ src/VIEW_CREATOR/ButtonFactory.java
@@ -1,4 +1,4 @@
-package Entities;
+package VIEW_CREATOR;
 
 import javafx.scene.control.Button;
 import javafx.scene.image.Image;
