+++ src/VIEW_CREATOR/LoadEventsViewFactory.java
@@ -1,0 +1,4 @@
+package VIEW_CREATOR;
+
+public class LoadEventsViewFactory {
+}
