+++ src/VIEW_CREATOR/LoadEventsViewModel.java
@@ -1,0 +1,4 @@
+package VIEW_CREATOR;
+
+public class LoadEventsViewModel {
+}
