+++ src/VIEW_CREATOR/LoadMapViewFactory.java
@@ -1,0 +1,4 @@
+package VIEW_CREATOR;
+
+public class LoadMapViewFactory {
+}
