+++ src/VIEW_CREATOR/LoadMapViewModel.java
@@ -1,0 +1,4 @@
+package VIEW_CREATOR;
+
+public class LoadMapViewModel {
+}
